+++ Runtime/Views/DotsViewBootstrap.cs
@@ -95,10 +95,14 @@ namespace Cuvara.DOTS.Views
             gameplay.AddSystemToUpdateList(commandBuffer);
 
             var view = world.GetOrCreateSystemManaged<ViewSystemGroup>();
+            var viewLifecycle = world.GetOrCreateSystemManaged<ViewLifecycleGroup>();
+            var viewSync = world.GetOrCreateSystemManaged<ViewTransformSyncGroup>();
             presentation.AddSystemToUpdateList(view);
-            view.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewSpawnSystem>());
-            view.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewDespawnSystem>());
-            view.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewTransformSyncSystem>());
+            view.AddSystemToUpdateList(viewLifecycle);
+            view.AddSystemToUpdateList(viewSync);
+            viewLifecycle.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewDespawnSystem>());
+            viewLifecycle.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewSpawnSystem>());
+            viewSync.AddSystemToUpdateList(world.GetOrCreateSystem<EntityViewTransformSyncSystem>());
 
             // Sorting is not automatic after a manual add: without this the UpdateAfter chain inside
             // each group is declared but not applied.